refactor: testCode 수정
