+++ service/product/src/main/java/jabaclass/product/application/dto/FileConfirmResponse.java
@@ -1,4 +1,4 @@
-package jabaclass.product.presentation.dto.response;
+package jabaclass.product.application.dto;
 
 import java.util.UUID;
